backend/requirements.txt
fastapi
uvicorn
python-dotenv
Pillow
PyMuPDF
google-generativeai
pytesseract
google-cloud-storage==2.16.0
google-auth==2.29.0
python-multipart
